Address PR review: printer package, no line width, docs
- Extract the doc renderer into internal/sql/ast/printer; TrackedBuffer
  embeds printer.Buffer and keeps the AST-facing layer (node dispatch,
  comment emission) in the ast package, since every node's Format method
  names TrackedBuffer and moving it would create an import cycle.
- Drop the 80-column line width: like gofmt, fmt no longer rewraps lines
  on its own. AttachComments records which printer-modeled boundaries the
  author broke at, and boundary() keeps those breaks; one-line statements
  stay on one line.
- Fold the comment-free path into the attachment path: formatStmt always
  goes through formatWithComments, retiring formatRaw/verifyFormatted.
- Cobra: Short is now "Format SQL queries"; Long drops the
  comments-never-deleted sentence and the engine-support paragraph.
- Docs: rewrite howto/fmt.md for the no-width model with a regenerated
  example; update cli.md; regenerate the sqlite endtoend golden.
