+++ internal/sql/ast/sub_link.go
@@ -54,12 +54,12 @@ func (n *SubLink) Format(buf *TrackedBuffer, d format.Dialect) {
 			buf.WriteString("(")
 		}
 	}
-	buf.group()
-	buf.indent()
-	buf.softline()
+	buf.Group()
+	buf.Indent()
+	buf.Softline()
 	buf.astFormat(n.Subselect, d)
-	buf.endIndent()
-	buf.softline()
-	buf.endGroup()
+	buf.EndIndent()
+	buf.Softline()
+	buf.EndGroup()
 	buf.WriteString(")")
 }
